+++ bin/commands/matchdatalogging.disable.js
@@ -13,14 +13,14 @@ class MatchDataLoggingCommand {
     sywac
       .number('--config <id>', {
         desc: 'Configuration ID. Mandatory if you have more than one configuration.',
-        group: 'Optional:',
-        required: false
+        group: 'Required:',
+        required: true
       })
       .string('--version <id>', {
         desc:
           "Version Number. It can also take the values 'PROD' or 'PRODUCTION' or 'STAGING'. If not provided, latest version is assumed.",
-        group: 'Optional:',
-        required: false
+        group: 'Required:',
+        required: true
       });
   }
   run(options) {