+++ bin/commands/matchdatalogging.js
@@ -13,20 +13,20 @@ class MatchDataLoggingCommand {
     sywac
       .number('--config <id>', {
         desc: 'Configuration ID. Mandatory if you have more than one configuration.',
-        group: 'Optional:',
-        required: false
+        group: 'Required:',
+        required: true
       })
       .string('--version <id>', {
         desc:
           "Version Number. It can also take the values 'PROD' or 'PRODUCTION' or 'STAGING'. If not provided, latest version is assumed.",
-        group: 'Optional:',
-        required: false
+        group: 'Required:',
+        required: true
       })
       .string('--policy <id>', {
         desc:
           'Policy ID. If provided, returns policy-level settings. If not provided, returns config-level settings.',
-        group: 'Optional:',
-        required: false
+        group: 'Required:',
+        required: true
       });
   }
   run(options) {