+++ bin/commands/matchdatalogging.modify.js
@@ -12,18 +12,20 @@ class MatchDataLoggingCommand {
   setup(sywac) {
     sywac
       .positional('<@path>', {
-        paramsDesc: 'The input file path.'
+        paramsDesc: 'The input file path.',
+        group: 'Required:',
+        required: true
       })
       .number('--config <id>', {
         desc: 'Configuration ID. Mandatory if you have more than one configuration.',
-        group: 'Optional:',
-        required: false
+        group: 'Required:',
+        required: true
       })
       .string('--version <id>', {
         desc:
           "Version Number. It can also take the values 'PROD' or 'PRODUCTION' or 'STAGING'. If not provided, latest version is assumed.",
-        group: 'Optional:',
-        required: false
+        group: 'Required:',
+        required: true
       })
       .check((argv, context) => {
         if (!argv['@path'].startsWith('@')) {